Restructure some navigation code to allow computing camera pose from node
